# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SELETIVO",
        "version": "0.1"
    },
    "paths": {
...
        "/api/foto-pessoa": {
            "get": {
                "tags": [
                    "FotoPessoa"
                ],
                "summary": "GET Foto Pessoa",
                "description": "GET Paginado",
                "operationId": "d00eae9085a87a12a72fe0d482ac5475",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/store-foto-pessoa/{pes_id}": {
            "post": {
                "tags": [
                    "FotoPessoa"
                ],
                "summary": "Armazenar foto vínculada a pessoa",
                "description": "Endpoint enviar foto de usuário, vinculando ao mesmo.",
                "operationId": "9400dab5fcf08c2383d769f3ad83295f",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "foto": {
                                        "description": "Arquivo de foto (imagem)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto enviada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Foto enviada com sucesso!"
                                        },
                                        "data": {
                                            "properties": {
                                                "documento_url": {
                                                    "type": "string",
                                                    "example": "https://seuservidor.com/uploads/documento.pdf"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-foto-pessoa/{pes_id}": {
            "post": {
                "tags": [
                    "FotoPessoa"
                ],
                "summary": "Atualizar foto de usuário",
                "description": "Endpoint atualizar foto de usuário, vinculando ao mesmo.",
                "operationId": "35b721210a6f7166d72fa851013e8b2d",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "foto"
                                ],
                                "properties": {
                                    "foto": {
                                        "description": "Arquivo da imagem",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto enviada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Foto enviada com sucesso!"
                                        },
                                        "data": {
                                            "properties": {
                                                "documento_url": {
                                                    "type": "string",
                                                    "example": "https://seuservidor.com/uploads/documento.pdf"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/show-foto-pessoa/{pes_id}": {
            "get": {
                "tags": [
                    "FotoPessoa"
                ],
                "summary": "Trazer foto do usuário com link temporário",
                "description": "Endpoint retornar link temporário da foto do usuário",
                "operationId": "19c7f0807e5ef32c1a56a657688073e7",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foto encontrada com sucesso!"
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-foto-pessoa/{pes_id}": {
            "delete": {
                "tags": [
                    "FotoPessoa"
                ],
                "summary": "Deleta Foto Pessoa",
                "description": "Remove Foto Pessoa",
                "operationId": "cbb92e683a532eaabf15bfae2c13cbda",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da Pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foto Pessoa Removida",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Foto Pessoa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}